# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pet Shop",
        "description": "Documentação da API Pet Shop",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pets": {
            "get": {
                "summary": "Listar pets",
                "operationId": "7b44ba1ad8a38f7865a2845fa9ebbfe3",
                "responses": {
                    "200": {
                        "description": "Lista de pets"
                    }
                }
            },
            "post": {
                "summary": "Cadastrar pet",
                "operationId": "9523721805529edb47b46332d56f2826",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Pet"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pet criado"
                    }
                }
            }
        },
        "/api/pets/{id}": {
            "get": {
                "summary": "Mostrar pet",
                "operationId": "25e25b195a59c4b940a33f9961ad6eed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pet encontrado"
                    }
                }
            },
            "put": {
                "summary": "Atualizar pet",
                "operationId": "652be7eab297dff87f093d849bdecb12",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Pet"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pet atualizado"
                    }
                }
            },
            "delete": {
                "summary": "Deletar pet",
                "operationId": "dab25348cae9362e48ade0e7cf34c465",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pet deletado"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "summary": "Listar usuários",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Lista de usuários"
                    }
                }
            },
            "post": {
                "summary": "Cadastrar usuário",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "summary": "Mostrar usuário",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado"
                    }
                }
            },
            "put": {
                "summary": "Atualizar usuário",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado"
                    }
                }
            },
            "delete": {
                "summary": "Deletar usuário",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário deletado"
                    }
                }
            }
        },
        "/api/users/{userId}/pets/{petId}": {
            "post": {
                "summary": "Adicionar pet ao usuário",
                "operationId": "16585a21b51403a2e9a675909b6ab9c4",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "petId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pet vinculado ao usuário"
                    }
                }
            },
            "delete": {
                "summary": "Remover pet do usuário",
                "operationId": "72368532250cd96cfa1740a1c403101e",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "petId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pet desvinculado do usuário"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Pet": {
                "title": "Pet",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "raca": {
                        "type": "string"
                    },
                    "idade": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}